object_detection_project/requirements.txt
 opencv-python-headless
tensorflow
mysql-connector-python
Flask

